+++ src/resource/mesh.cpp
@@ -31,6 +31,7 @@
 #   include "resource/data_compiler.h"
 #   include "resource/mesh.h"
 #   include "resource/mesh_fbx.h"
+#   include "resource/mesh_gltf.h"
 #   include "resource/mesh_obj.h"
 #   include "resource/mesh_resource.h"
 #   include <bx/error.h>
@@ -542,6 +543,14 @@ namespace mesh
 			RETURN_IF_FILE_MISSING(MESH, source.c_str(), opts);
 			if (str_has_suffix_case(source.c_str(), ".obj"))
 				return crown::obj::parse(m, source.c_str(), opts);
+			if (str_has_suffix_case(source.c_str(), ".gltf")
+				|| str_has_suffix_case(source.c_str(), ".glb")
+				)
+				return crown::gltf::parse(m
+					, source.c_str()
+					, json_object::has(obj, "rigid_attachments") ? obj["rigid_attachments"] : NULL
+					, opts
+					);
 
 			Buffer fbx_buf = opts.read(source.c_str());
 			return fbx::parse(m, fbx_buf, opts);