+++ docs/changelog.rst
@@ -16,7 +16,7 @@ Changelog
 * Tools: searching large projects in the Project Browser is now much quicker.
 * Tools: optimized vertical space usage in the editor windows by moving the menubar into the titlebar.
 * Tools: increased icons density in the Project Browser by removing excessive padding and margins.
-* Tools: property grids' widgets now better adapt when the containing panel is resized horizontally.
+* Tools: added support for importing glTF 2.0 scenes in ``.gltf`` and ``.glb`` formats.
 * Runtime: added support for MP3 sound files.
 * Runtime: added triplanar mapping support to the mesh shader (+TRIPLANAR and +TRIPLANAR_LOCAL).
 * Runtime: added a new Vignette post-processing effect.